refactor: Components, Layers 폴더 구조화 [#98]
- Backlog 폴더 Layers 로 이동
